RAPI: Workaround twisted failure by early import rpc
This could probably solved in a nicer way, but it seems that somehow twisted
gets imported before our rpc which overrides the default twisted reactor; or
maybe it gets imported twice or such. Anyway, importing rpc early makes the rpc
layer work.

Lest this is taken as a random thing, this can be consistently
reproduced and traces of imports are not very clear of why this happens
- nor why gnt-os works, despite having a very similar import list and
not importing rpc.

Reviewed-by: amishchenko,imsnah