Implement network reconfiguration for drbd8
This patch implements a simple network reconfiguration method for DRBD8
devices: it currently only allows changing of the multi-master flag.

The method can only be called when the device is attached, and it will
not work for (primary and diskless) devices, as then the shutdown will
fail; the consistency needs to be checked before attempting the
reconfiguration.

Reviewed-by: imsnah