AddNode: Check for node existance
In the "new world" we'll need to setup ganeti-noded via ssh on the node
before calling the AddNode opcode. Before doing it we'll check that the
node is not already in the cluster, if --readd was not passed. This
guarantees we're not going to restart ganeti-noded on a running node.

This patch also incidentally fixes a non-style-guide conformant
docstring.

Reviewed-by: iustinp